Delete old graph builder.
