+++ crates/project-graph/src/lib.rs
@@ -1,16 +1,7 @@
-mod project_events;
 mod project_graph;
-mod project_graph_builder;
-mod project_graph_cache;
 mod project_graph_error;
-mod project_graph_hash;
 mod project_matcher;
-mod projects_locator;
 
-pub use project_events::*;
 pub use project_graph::*;
-pub use project_graph_builder::*;
-pub use project_graph_cache::*;
 pub use project_graph_error::*;
 pub use project_matcher::*;
-pub use projects_locator::*;
