--- crates/project-graph/src/project_graph_cache.rs
@@ -1,11 +1,0 @@
-use moon_cache::cache_item;
-use moon_common::path::WorkspaceRelativePathBuf;
-use moon_common::Id;
-use rustc_hash::FxHashMap;
-
-cache_item!(
-    pub struct ProjectsCacheState {
-        pub last_hash: String,
-        pub projects: FxHashMap<Id, WorkspaceRelativePathBuf>,
-    }
-);
