+++ crates/workspace/src/lib.rs
@@ -1,9 +1,12 @@
+mod project_build_data;
 mod projects_locator;
 mod repo_type;
 mod workspace_builder;
 mod workspace_builder_error;
 mod workspace_cache;
 
+pub use project_build_data::*;
 pub use repo_type::*;
 pub use workspace_builder::*;
 pub use workspace_builder_error::*;
+pub use workspace_cache::*;
