+++ crates/workspace/src/project_build_data.rs
@@ -1,9 +1,40 @@
 use moon_common::path::WorkspaceRelativePathBuf;
 use moon_common::Id;
-use moon_config::{DependencyConfig, ProjectsAliasesList, ProjectsSourcesList, TaskConfig};
+use moon_config::{
+    DependencyConfig, ProjectConfig, ProjectsAliasesList, ProjectsSourcesList, TaskConfig,
+};
+use petgraph::graph::NodeIndex;
 use rustc_hash::FxHashMap;
+use serde::{Deserialize, Serialize};
 use starbase_events::Event;
 use std::path::PathBuf;
+
+#[derive(Clone, Debug, Default, Deserialize, PartialEq, Serialize)]
+#[serde(default)]
+pub struct ProjectBuildData {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub alias: Option<String>,
+
+    #[serde(skip)]
+    pub config: Option<ProjectConfig>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub node_index: Option<NodeIndex>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub original_id: Option<Id>,
+
+    pub source: WorkspaceRelativePathBuf,
+}
+
+impl ProjectBuildData {
+    pub fn from_source(source: &str) -> Self {
+        Self {
+            source: source.into(),
+            ..Default::default()
+        }
+    }
+}
 
 // Extend the project graph with additional information.
